+++ NEWS.md
@@ -1,11 +1,18 @@
 # inlabru (development version)
 
-## Feaures
+## Features
   
 * Add `bru_get_mapper` generic, and associated methods for `inla.spde` and
   `inla.rgeneric` objects. This allows `inlabru` to automatically extract
   the appropriate `bru_mapper` object for each model component, and can be used
   as a hook by external packages implementing new INLA object classes.
+  
+* Add a `weights` argument for `like()`, for likelihood-specific log-likelihood
+  weights, passed on to the `INLA::inla()` weights argument. Evaluated in the
+  data context.
+  
+* New vignettes on the `bru_mapper` system, `component` definitions,
+  and `prediction_scores`
 
 ## Bug fixes